requirements.txt
tensorflow-gpu==1.8.0
tqdm
ujson
spacy
click
numpy